requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
reportlab==4.0.8
fpdf2==2.7.6